--- product-service/src/main/java/com/lotfresh/productservice/domain/category/CategoryRepository.java
@@ -1,6 +1,0 @@
-package com.lotfresh.productservice.domain.category;
-
-import com.lotfresh.productservice.domain.category.custom.CategoryRepositoryCustom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long>, CategoryRepositoryCustom {}
