+++ product-service/src/main/java/com/lotfresh/productservice/domain/category/api/controller/request/CategoryCreateRequest.java
@@ -1,6 +1,6 @@
-package com.lotfresh.productservice.api.controller.request;
+package com.lotfresh.productservice.domain.category.api.controller.request;
 
-import com.lotfresh.productservice.domain.category.Category;
+import com.lotfresh.productservice.domain.category.entity.Category;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
