+++ product-service/src/main/java/com/lotfresh/productservice/domain/category/api/service/CategoryService.java
@@ -1,6 +1,6 @@
-package com.lotfresh.productservice.api.service.category;
+package com.lotfresh.productservice.domain.category.api.service;
 
-import com.lotfresh.productservice.domain.category.CategoryRepository;
+import com.lotfresh.productservice.domain.category.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
