--- product-service/src/main/java/com/lotfresh/productservice/domain/category/custom/CategoryRepositoryCustom.java
@@ -1,3 +1,0 @@
-package com.lotfresh.productservice.domain.category.custom;
-
-public interface CategoryRepositoryCustom {}
