+++ product-service/src/main/java/com/lotfresh/productservice/domain/category/repository/CategoryRepository.java
@@ -1,0 +1,8 @@
+package com.lotfresh.productservice.domain.category.repository;
+
+import com.lotfresh.productservice.domain.category.entity.Category;
+import com.lotfresh.productservice.domain.category.repository.custom.CategoryRepositoryCustom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> , CategoryRepositoryCustom {
+}
