+++ product-service/src/main/java/com/lotfresh/productservice/domain/category/repository/custom/CategoryRepositoryCustom.java
@@ -1,0 +1,3 @@
+package com.lotfresh.productservice.domain.category.repository.custom;
+
+public interface CategoryRepositoryCustom {}
