+++ product-service/src/main/java/com/lotfresh/productservice/domain/category/repository/custom/CategoryRepositoryImpl.java
@@ -1,9 +1,9 @@
-package com.lotfresh.productservice.domain.category.custom;
+package com.lotfresh.productservice.domain.category.repository.custom;
 
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public class CategoryRepositoryImpl implements CategoryRepositoryCustom {
+public class CategoryRepositoryImpl implements CategoryRepositoryCustom{
     private final JPAQueryFactory query;
 }
