+++ product-service/src/main/java/com/lotfresh/productservice/domain/discount/repository/DiscountRepository.java
@@ -1,5 +1,6 @@
-package com.lotfresh.productservice.domain.discount;
+package com.lotfresh.productservice.domain.discount.repository;
 
+import com.lotfresh.productservice.domain.discount.entity.Discount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DiscountRepository extends JpaRepository<Discount, Long> {}
