+++ product-service/src/main/java/com/lotfresh/productservice/domain/product/api/service/ProductService.java
@@ -1,6 +1,6 @@
-package com.lotfresh.productservice.api.service.product;
+package com.lotfresh.productservice.domain.product.api.service;
 
-import com.lotfresh.productservice.domain.product.ProductRepository;
+import com.lotfresh.productservice.domain.product.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
