+++ product-service/src/main/java/com/lotfresh/productservice/domain/product/entity/Product.java
@@ -1,7 +1,7 @@
-package com.lotfresh.productservice.domain.product;
+package com.lotfresh.productservice.domain.product.entity;
 
-import com.lotfresh.productservice.domain.BaseEntity;
-import com.lotfresh.productservice.domain.category.Category;
+import com.lotfresh.productservice.common.BaseEntity;
+import com.lotfresh.productservice.domain.category.entity.Category;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
