+++ product-service/src/main/java/com/lotfresh/productservice/domain/product/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package com.lotfresh.productservice.domain.product;
+package com.lotfresh.productservice.domain.product.repository;
 
+import com.lotfresh.productservice.domain.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {}
